@@ -4,20 +4,17 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   search_app_bar:
     path: ../
-  rxdart:
-    git:
-      url: https://github.com/ReactiveX/rxdart
-      ref: master
-  diacritic: ^0.1.1
-  cupertino_icons: ^0.1.2
-  dio: ^3.0.10
+  rxdart: ^0.27.1
+  diacritic: ^0.1.3
+  cupertino_icons: ^0.1.3
+  dio: ^4.0.0
   json_annotation:
 
 dev_dependencies:
